pyproject.toml
[tool.poetry]
name = "donut"
version = "0.1.0"
description = ""
authors = ["Toshiyuki Haruhara <toshiyuki1998@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.43.3"
torch = "^2.4.0"
torchvision = "^0.19.0"
torchaudio = "^2.4.0"
datasets = "^2.20.0"
pillow = "^10.4.0"
sentencepiece = "^0.2.0"
protobuf = "^5.27.2"
ipykernel = "^6.29.5"
openpyxl = "^3.1.5"
pandas = "^2.2.2"

pdf2image = "^1.17.0"
xlsx2html = "^0.6.1"
pdfkit = "^1.0.0"
sudachidict-small = "^20240716"
faker = "^26.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
